MInor code fixtures
