+++ pkg/admin/zkclient.go
@@ -1076,3 +1076,31 @@ func updateConfig(
 	configMap["config"] = configKVMap
 	return updatedKeys, nil
 }
+
+func (c *ZKAdminClient) GetAllTopicsMetadata(
+	ctx context.Context,
+) (*kafka.MetadataResponse, error) {
+	topicsInfo, err := c.GetTopics(ctx, nil, false)
+	if err != nil {
+		return nil, fmt.Errorf("Error fetching all topics info: %+v", err)
+	}
+	log.Debugf("topicsInfo: %v", topicsInfo)
+
+	allTopics := []string{}
+	for _, topicInfo := range topicsInfo {
+		allTopics = append(allTopics, topicInfo.Name)
+	}
+
+	client := c.GetConnector().KafkaClient
+	req := kafka.MetadataRequest{
+		Topics: allTopics,
+	}
+
+	log.Debugf("Metadata request: %+v", req)
+	metadata, err := client.Metadata(ctx, &req)
+	if err != nil {
+		return nil, fmt.Errorf("Error fetching all topics metadata: %+v", err)
+	}
+
+	return metadata, nil
+}
